package.json: save file like npm would
Minimize the diffs when saving package.json by leveraging the original
npm method. It determines the pre-existing indentation and newline
style used in the file and re-applies it to the updated content
(with limitations, cf. [1]).

An alternative would have been to look for solutions which operate
comparable to the manipulation of binary files, exchanging only the
differing bits, but this seems much more involved (i.e. risky) and
possibly in vain if npm would - sooner or later - overwrite them
anyway.

Resolves #15

[1]: https://github.com/npm/read-package-json-fast#indentation